Reformulate the entry to the outer loop of BLOCK3 to avoid spurious early termination.
In the previous version of the initilization, we started by solving
the unconstrained problem and projecting the solution into the x>=0
space. The first operation in the other loop is to check the Lagrange
multipliers associated with the constrained coefficients and free any
whose nonzeroness would reduce the residual. In certain cases where a
very small penalty term was used, it could happen that all elements of y
were nonnegative, causing the outer loop to terminate with a truncated
solution that did not correspond to the minimum in the subspace of
the free coefficients. Starting with the null vector instead of the
truncated solution to the unconstrained problem keeps the state of the
factorization consistent with the rest of the outer loop, and makes this
behavior go away.



git-svn-id: http://code.icecube.wisc.edu/svn/projects/photospline/trunk@90434 16731396-06f5-0310-8873-f7f720988828